// client/src/features/RegistrationByPhone/model/types/RegistrationByPhoneValidateErrorTypes.ts
export enum RegistrationByPhoneValidateErrorTypes {
    // Ошибки до запроса на сервер

    // номер телефона
    EMPTY_PHONE_ERROR = 'EMPTY_PHONE_ERROR',
    FORMAT_PHONE_ERROR = 'FORMAT_PHONE_ERROR',

    // смс код
    EMPTY_CODE_ERROR = 'EMPTY_CODE_ERROR',
    FORMAT_CODE_ERROR = 'FORMAT_CODE_ERROR',

    // почта
    FORMAT_EMAIL_ERROR = 'FORMAT_EMAIL_ERROR',

    // пароль
    EMPTY_PASSWORD_ERROR = 'EMPTY_PASSWORD_ERROR',
    SYMBOLS_PASSWORD_ERROR = 'SYMBOLS_PASSWORD_ERROR',
    LENGTH_PASSWORD_ERROR = 'LENGTH_PASSWORD_ERROR',

    // пароль подтверждения
    EMPTY_PASSWORD_CONFIRM_ERROR = 'EMPTY_PASSWORD_CONFIRM_ERROR',
    INVALID_PASSWORD_CONFIRM_ERROR = 'INVALID_PASSWORD_CONFIRM_ERROR',
}